@@ -15,50 +15,28 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+11
+version: 1.0.0+12
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.5.0 <3.0.0"
+  sdk: ">=2.15.1 <3.0.0"
+  flutter: ">=2.8.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  stacked: ^2.3.12
+  stacked_services: ^0.8.26
+  dropdown_search: ^5.0.2
+  url_launcher: ^6.1.4
+  http: ^0.13.4
+  result_type: ^0.1.0
 
-
-  http: ^0.13.3
-  xml: ^5.3.0
-  google_maps_flutter: ^2.0.10
-  maps_launcher: ^2.0.1
-  shared_preferences: ^2.0.8
-  url_launcher: ^6.0.12
-  flutter_local_notifications: ^8.2.0
-  background_fetch: ^1.0.1
-  searchable_dropdown: ^1.1.3
-  dropdown_search: ^1.0.4 #^0.6.1 #- can't use until they fix keyboard overlapping bottom sheet search bar
-   # path: ./local-plugins/dropdown_search-0.6.1-local
-  sqflite: ^2.0.0+4
-  introduction_screen: ^2.1.0
-  intl: ^0.17.0
-  flutter_keyboard_visibility: ^5.0.3
-  pull_to_refresh: ^2.0.0
-  provider: ^5.0.0
-  firebase_messaging: ^10.0.7
-  flutter_layout_grid: ^1.0.3
-  flutter_native_splash: ^1.2.4
-  flutter_launcher_icons: ^0.9.2
-
-  device_preview: ^0.8.0
-
-  cupertino_icons: ^1.0.3
 
 dev_dependencies:
-  test: any
-  flutter_driver:
-    sdk: flutter
-  mockito: ^5.0.16
-  build_runner: ^2.1.2
-  flutter_lints: ^1.0.4
+  flutter_native_splash: ^2.2.3+1
+  json_annotation: ^4.5.0
+  build_runner: ^2.1.11
+  stacked_generator: ^0.7.9+1
 
 
 # flutter pub run flutter_native_splash:create
@@ -66,6 +44,7 @@ flutter_native_splash:
   color: '#d96d19'
   fullscreen: true
   image: 'assets/splash/Perth Fuel Prices Splash Screen.png'
+  android12:
 
 
 # flutter pub run flutter_launcher_icons:main
